Create paddles

+++ src/components/position.rs
@@ -3,12 +3,12 @@ use super::component_prelude::*;
 #[derive(Component)]
 #[storage(VecStorage)]
 pub struct Position {
-    pub x: u16,
-    pub y: u16,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Position {
-    pub fn new(x: u16, y: u16) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         Self { x, y }
     }
 }
